Refactor: DRY utility lambda functions
