+++ packages/devops/scripts/lambda/deleteDeployment.py
@@ -4,28 +4,13 @@ import datetime
 import re
 import asyncio
 import functools
+from utilities import *
 
 ec2 = boto3.resource('ec2')
 ec = boto3.client('ec2')
 iam = boto3.client('iam')
 route53 = boto3.client('route53')
 
-def get_tag(instance, tag_name):
-    try:
-        tag_value = [
-            t.get('Value') for t in instance['Tags']
-            if t['Key'] == tag_name][0]
-    except IndexError:
-        tag_value = ''
-    return tag_value
-
-def get_instances(filters):
-    reservations = ec.describe_instances(
-        Filters=filters
-    ).get(
-        'Reservations', []
-    )
-    return reservations[0]['Instances']
 
 def build_record_set_deletion(domain, subdomain, stage, ip_address):
     if (subdomain == ''):
